fix: always use int64 for full-to-swa / hisparse index translation
full_to_swa_index_mapping and full_to_hisparse_device_index_mapping are
both allocated as int64, but translate_loc_from_full_to_swa and
translate_loc_to_hisparse_device down-cast the indexed result to int32, so
callers indexing the mapping directly saw int64 while callers going
through these methods saw int32. Drop the casts so the natural int64
result flows through consistently.

+++ python/sglang/srt/mem_cache/deepseek_v4_memory_pool.py
@@ -531,7 +531,7 @@ class DeepSeekV4TokenToKVPool(BaseSWAKVPool):
     def translate_loc_from_full_to_swa(self, kv_indices: torch.Tensor):
         assert self.full_to_swa_index_mapping is not None
 
-        return self.full_to_swa_index_mapping[kv_indices].to(torch.int32)
+        return self.full_to_swa_index_mapping[kv_indices]
 
     def get_cached_swa_loc(self, raw_loc: torch.Tensor, layer_id: int) -> torch.Tensor:
         if self._should_cache_swa:
